+++ src/app/page/phieu-kiem-ke/list-phieu-kiem-ke/list-phieu-kiem-ke.component.ts
@@ -4,6 +4,8 @@ import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import { PhieuKiemKeService } from 'src/app/services/phieu-kiem-ke.service';
 import { Title } from '@angular/platform-browser';
+import { User } from 'src/app/models/user';
+import { AuthenticationService } from 'src/app/services/authentication.service';
 
 
 @Component({
@@ -29,18 +31,21 @@ export class ListPhieuKiemKeComponent implements OnInit {
     "contentMsg": ''
   };
   @ViewChild(MatPaginator) paginator!: MatPaginator;
+  currentUser!: User
   constructor(
     private router: Router,
     private titleService: Title,
 
-    private phieuKiemKeService: PhieuKiemKeService
+    private phieuKiemKeService: PhieuKiemKeService,
+    private authenService: AuthenticationService
   ) { 
     this.titleService.setTitle('Danh sách phiếu kiểm kê');
     if (this.router.getCurrentNavigation()?.extras.state) {
       this.alerMsg['showMsg'] = true;
       this.alerMsg['typeMsg'] = this.router.getCurrentNavigation()?.extras.state?.typeMsg;
       this.alerMsg['contentMsg'] = this.router.getCurrentNavigation()?.extras.state?.contentMsg;
-    }
+    };
+    this.currentUser = this.authenService.userValue;
   }
 
   ngOnInit(): void {
@@ -67,22 +72,30 @@ export class ListPhieuKiemKeComponent implements OnInit {
 
 
   async deletePhieuKiemKe(id: string) {
-    if(confirm("Bạn có chắc muốn delete ?")) {
-      const deletePhieuKiemKe = this.phieuKiemKeService.deletePhieuKiemKe(id).toPromise().then(
-        () => {
-          this.alerMsg['showMsg'] = true;
-          this.alerMsg['typeMsg'] = 'success';
-          this.alerMsg['contentMsg'] = 'Delete success';
-        },
-        () => {
-          this.alerMsg['showMsg'] = true;
-          this.alerMsg['typeMsg'] = 'danger';
-          this.alerMsg['contentMsg'] = 'Delete failed!';
-        }
-      );
-  
-      await Promise.all([deletePhieuKiemKe]);
-      this.ngOnInit();
+    if(this.currentUser.role != "Admin" && this.currentUser.role != "NhaBep")
+    {
+      this.alerMsg['showMsg'] = true;
+      this.alerMsg['typeMsg'] = 'danger';
+      this.alerMsg['contentMsg'] = 'Không được phép xóa';
+    }
+    else{
+      if(confirm("Bạn có chắc muốn delete ?")) {
+        const deletePhieuKiemKe = this.phieuKiemKeService.deletePhieuKiemKe(id).toPromise().then(
+          () => {
+            this.alerMsg['showMsg'] = true;
+            this.alerMsg['typeMsg'] = 'success';
+            this.alerMsg['contentMsg'] = 'Delete success';
+          },
+          () => {
+            this.alerMsg['showMsg'] = true;
+            this.alerMsg['typeMsg'] = 'danger';
+            this.alerMsg['contentMsg'] = 'Delete failed!';
+          }
+        );
+    
+        await Promise.all([deletePhieuKiemKe]);
+        this.ngOnInit();
+      }
     }
   }
   applyFilter(event: Event) {
